Adding component UI with JSX
In JSX, Use 'className' instead of 'class' as class is a reserved keyword in JS
If you want to return an element on its own line, you have to use a pair of parenthesis.
You should return only one parent element, we can't return two sibling elements without wrapping them in a parent element.
Capital letters is the naming convention for React components